Update minimization logic & directory structure
